DAO-203 Move components/policies -> pages/components/policies
