+++ src/pages/components/policies/policies.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.module.scss';
+@import '../../../styles/variables.module.scss';
 
 .emptyState {
   color: $secondary-color;
