+++ src/pages/components/policies/policy-list.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.module';
+@import '../../../styles/variables.module';
 
 .policyList {
   padding: 0;
