+++ src/pages/components/policies/search-form.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.module.scss';
+@import '../../../styles/variables.module.scss';
 
 .searchForm {
   display: flex;
