+++ src/pages/components/policies/search-form.tsx
@@ -1,7 +1,7 @@
 import { FormEventHandler } from 'react';
-import Input from '../input';
-import SearchIcon from '../icons/search-icon';
-import CloseIcon from '../icons/close-icon';
+import Input from '../../../components/input';
+import SearchIcon from '../../../components/icons/search-icon';
+import CloseIcon from '../../../components/icons/close-icon';
 import styles from './search-form.module.scss';
 
 interface Props {
