+++ src/pages/my-policies/my-policies.tsx
@@ -3,8 +3,8 @@ import { useHistory } from 'react-router';
 import Layout from '../../components/layout';
 import BorderedBox, { Header } from '../../components/bordered-box';
 import RadioButton from '../../components/radio-button';
-import Policies, { Filter } from '../../components/policies';
-import SearchForm from '../../components/policies/search-form';
+import Policies, { Filter } from '../components/policies';
+import SearchForm from '../components/policies/search-form';
 import { useQueryParams } from '../../utils';
 import styles from './my-policies.module.scss';
